etl_pipeline/requirements.txt
dagster
dagster-webserver
minio
requests
python-dotenv
pandas
clickhouse-driver